Render lines

+++ chip_8.h
@@ -19,7 +19,7 @@ class CHIP_8 {
         uint16_t _pc;           // 16 bit, store currently executing address
 
         uint8_t _sp;            // 8 bit stack pointer to topmost level of stack
-        uint8_t _stack[16];     // 16 levels of stack
+        uint16_t _stack[16];     // 16 levels of stack
 
         std::array<int, 80> _fontset;
